Move the remaining wire type modules into probe-rs-rpc
