+++ probe-rs-rpc/src/lib.rs
@@ -129,9 +129,12 @@ pub type NoResponse = RpcResult<()>;
 pub mod breakpoints;
 pub mod chip;
 pub mod core_ops;
+pub mod debug_vars;
+pub mod disassemble;
 pub mod file;
 pub mod flash;
 pub mod format;
+pub mod info;
 pub mod memory;
 pub mod monitor;
 pub mod probe;
@@ -140,3 +143,5 @@ pub mod resume;
 pub mod rtt_client;
 pub mod rtt_config;
 pub mod semihosting_options;
+pub mod stack_trace;
+pub mod test;
