+++ probe-rs-tools/src/bin/probe-rs/rpc/functions/debug_vars.rs
@@ -1,92 +1,14 @@
-use crate::rpc::{
-    Key, Session,
-    functions::{RpcContext, RpcResult, convert::lift},
-};
 use postcard_rpc::header::VarHeader;
-use postcard_schema::Schema;
 use probe_rs_debug::{
     DebugInfo, DebugRegisters, ObjectRef, StackFrameInfo, Variable, VariableCache, VariableName,
 };
-use serde::{Deserialize, Serialize};
+pub use probe_rs_rpc::debug_vars::{
+    ClearCoreDebugStateRequest, EvaluateRequest, EvaluateResponse, LoadSvdRequest, LoadSvdResponse,
+    ScopesRequest, ScopesResponse, SetVariableRequest, SetVariableResult, VariablesRequest,
+    VariablesResponse, WireEvaluateResponse, WireScope, WireSetVariableResponse, WireVariable,
+};
 
-#[derive(Serialize, Deserialize, Schema)]
-pub struct ScopesRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    pub frame_id: u32,
-}
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireScope {
-    pub name: String,
-    pub presentation_hint: Option<String>,
-    pub variables_reference: i64,
-    pub expensive: bool,
-    pub line: Option<i64>,
-    pub column: Option<i64>,
-}
-
-pub type ScopesResponse = crate::rpc::functions::RpcResult<Vec<WireScope>>;
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct VariablesRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    pub variables_reference: u32,
-    pub filter: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct ClearCoreDebugStateRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-}
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct LoadSvdRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    /// Server-side path to the CMSIS-SVD file (the client uploads it via the
-    /// temp-file endpoints, then passes the resulting path here), or `None`
-    /// to remove the core's current SVD state.
-    pub path: Option<String>,
-}
-
-pub type LoadSvdResponse = crate::rpc::functions::RpcResult<()>;
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct EvaluateRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    pub frame_id: Option<u32>,
-    pub expression: String,
-}
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireEvaluateResponse {
-    pub result: String,
-    pub type_: Option<String>,
-    pub variables_reference: i64,
-    pub named_variables: Option<i64>,
-    pub indexed_variables: Option<i64>,
-    pub memory_reference: Option<String>,
-}
-
-pub type EvaluateResponse = crate::rpc::functions::RpcResult<WireEvaluateResponse>;
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireVariable {
-    pub name: String,
-    pub evaluate_name: Option<String>,
-    pub memory_reference: Option<String>,
-    pub indexed_variables: Option<i64>,
-    pub named_variables: Option<i64>,
-    pub type_: Option<String>,
-    pub value: String,
-    pub variables_reference: i64,
-}
-
-pub type VariablesResponse = crate::rpc::functions::RpcResult<Vec<WireVariable>>;
+use crate::rpc::functions::{RpcContext, RpcResult, convert::lift};
 
 /// Mirrors `request_helpers::get_variable_reference` for the server-side path.
 fn variable_reference(parent: &Variable, cache: &VariableCache) -> (ObjectRef, i64, i64) {
@@ -338,7 +260,7 @@ pub async fn clear_core_debug_state(
     ctx: &mut RpcContext,
     _header: VarHeader,
     request: ClearCoreDebugStateRequest,
-) -> crate::rpc::functions::RpcResult<()> {
+) -> RpcResult<()> {
     let states = ctx.debug_states();
     let mut guard = states.lock().await;
     if let Some(state) = guard.get_mut(&request.sessid) {
@@ -373,27 +295,6 @@ pub async fn load_svd(
     })
     .await
 }
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct SetVariableRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    pub parent_key: i64,
-    pub name: String,
-    pub value: String,
-}
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireSetVariableResponse {
-    pub value: String,
-    pub type_: Option<String>,
-    pub variables_reference: i64,
-    pub named_variables: Option<i64>,
-    pub indexed_variables: Option<i64>,
-    pub memory_reference: Option<String>,
-}
-
-pub type SetVariableResult = RpcResult<WireSetVariableResponse>;
 
 /// Set a local/static variable's value server-side. The client only relays
 /// the parent key, name, and new value. Searches per-frame locals then
