+++ probe-rs-tools/src/bin/probe-rs/rpc/functions/disassemble.rs
@@ -1,41 +1,10 @@
-use postcard_schema::Schema;
-use serde::{Deserialize, Serialize};
-
-use crate::rpc::{
-    Key, Session,
-    functions::{RpcContext, RpcResult, convert::lift},
-};
-
-#[derive(Serialize, Deserialize, Schema)]
-pub struct DisassembleRequest {
-    pub sessid: Key<Session>,
-    pub core: u32,
-    pub memory_reference: u64,
-    pub byte_offset: i64,
-    pub instruction_offset: i64,
-    pub instruction_count: i64,
-}
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireSource {
-    pub name: Option<String>,
-    pub path: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct WireDisassembledInstruction {
-    pub address: String,
-    pub column: Option<i64>,
-    pub instruction: String,
-    pub instruction_bytes: Option<String>,
-    pub line: Option<i64>,
-    pub location: Option<WireSource>,
-}
-
-pub type DisassembleResponse = RpcResult<Vec<WireDisassembledInstruction>>;
-
 use crate::cmd::dap_server::debug_adapter::dap::request_helpers::disassemble_target_memory;
 use postcard_rpc::header::VarHeader;
+pub use probe_rs_rpc::disassemble::{
+    DisassembleRequest, DisassembleResponse, WireDisassembledInstruction, WireSource,
+};
+
+use crate::rpc::functions::{RpcContext, convert::lift};
 
 /// Disassemble target memory server-side, running the capstone disassembly
 /// (shared with the local path via `disassemble_target_memory`) against the
